Move repl to separate file
